+++ Taller/Anglesg.h
@@ -1,8 +1,8 @@
 #ifndef ANGLESG_H_INCLUDED
 #define ANGLESG_H_INCLUDED
 
-void anglesg(long double Alpha1, long double Alpha2, long double Alpha3, long double Delta1, long double Delta2, long double Delta3,
-                           long double JD1[3], long double JD2[3], long double JD3[3],
-                           long double RS1[3], long double RS2[3], long double RS3[3]);
+void anglesg( double Alpha1,  double Alpha2,  double Alpha3,  double Delta1,  double Delta2,  double Delta3,
+                            double JD1[3],  double JD2[3],  double JD3[3],
+                            double RS1[3],  double RS2[3],  double RS3[3]);
 
 #endif // ANGLESG_H_INCLUDED
