+++ Taller/EqnEquinox.c
@@ -3,9 +3,9 @@
 #include "MeanObliquity.h"
 #include "math.h"
 
-long double eqnEquinox (long double fechaJulianaModificada )
+ double eqnEquinox ( double fechaJulianaModificada )
 {
-    long double vectAngulos[2] = {0.0, 0.0};
+     double vectAngulos[2] = {0.0, 0.0};
     NutAngles(fechaJulianaModificada, vectAngulos);
     return vectAngulos[0] * cos(MeanObliquity(fechaJulianaModificada));
 }
