+++ Taller/Frac.h
@@ -1,6 +1,6 @@
 #ifndef FRAC_H_INCLUDED
 #define FRAC_H_INCLUDED
 
-long double Frac(long double x);
+ double Frac( double x);
 
 #endif // FRAC_H_INCLUDED
