+++ Taller/R_y.h
@@ -1,7 +1,7 @@
 #ifndef R_Y_H_INCLUDED
 #define R_Y_H_INCLUDED
 
-void R_y(long double angle, long double matrizRotada[3][3]);
+void R_y( double angle,  double matrizRotada[3][3]);
 
 
 #endif // R_Y_H_INCLUDED
